+++ app/Http/Controllers/MediaController.php
@@ -13,6 +13,7 @@ use App\Models\Media;
 use App\Models\MediaType;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 class MediaController extends Controller
 {
 
@@ -23,9 +24,12 @@ class MediaController extends Controller
      */
     public function index(Request $request)
     {
-        $medias = Media::orderBy('id','DESC')->paginate(5);
-        return view('media.index',compact('medias'))
-            ->with('i', ($request->input('page', 1) - 1) * 5);
+        if (Auth::check()) {
+            $medias = Media::orderBy('id', 'DESC')->paginate(5);
+            return view('media.index', compact('medias'))
+                ->with('i', ($request->input('page', 1) - 1) * 5);
+        } else
+            return view('errors.permission');
     }
 
     /**
@@ -35,10 +39,11 @@ class MediaController extends Controller
      */
     public function create()
     {
+        if (Auth::check()) {
 //        $types = MediaType::lists('name','id');
-        $locations = Location::lists('name','id');
-        $types = MediaType::lists('name', 'id');
-        return view('media.create', compact('id', 'types'));
+            $locations = Location::lists('name', 'id');
+            $types = MediaType::lists('name', 'id');
+            return view('media.create', compact('id', 'types'));
 //        $models = $->models();
 //        return Response::eloquent($models->get(['id','name']));
 //        return View::make('media.create')->with('locations', $locations)->with('types',$types);
@@ -46,6 +51,8 @@ class MediaController extends Controller
 //            ->with('locations', Location::orderBy('id', 'asc')->lists('name','id'))
 //            ->with('types', MediaType::orderBy('id', 'asc')->lists('name','id'));
 //        return view('media.create',compact('types','locations'));
+        } else
+            return view('errors.permission');
     }
 
     /**
@@ -56,33 +63,36 @@ class MediaController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'source' => 'required',
-            'address' => 'required',
-            'type_id' => 'required',
-            'user_id' => 'required',
-        ]);
+        if (Auth::check()) {
+            $this->validate($request, [
+                'source' => 'required',
+                'address' => 'required',
+                'type_id' => 'required',
+                'user_id' => 'required',
+            ]);
 
-        $file = $request->file('file_name');
-        $fileName = $file->getClientOriginalName();
-        $filePath = $file->getRealPath();
-        $destinationPath = 'uploads';
-        $file->move($destinationPath,$file->getClientOriginalName());
+            $file = $request->file('file_name');
+            $fileName = $file->getClientOriginalName();
+            $filePath = $file->getRealPath();
+            $destinationPath = 'uploads';
+            $file->move($destinationPath, $file->getClientOriginalName());
 
-        DB::table('media')->insert(
-            ['source' => $request->source,
-            'address' => $request->address,
-            'filePath' => $filePath,
-            'created_at' => new \DateTime(),
-            'updated_at' => new \DateTime(),
-            'name' => $request->name,
-            'type_id' => $request->type_id,
-            'user_id' => $request->user_id,]
-        );
+            DB::table('media')->insert(
+                ['source' => $request->source,
+                    'address' => $request->address,
+                    'filePath' => $filePath,
+                    'created_at' => new \DateTime(),
+                    'updated_at' => new \DateTime(),
+                    'name' => $request->name,
+                    'type_id' => $request->type_id,
+                    'user_id' => $request->user_id,]
+            );
 
 //        Media::create($request->all());
-        return redirect()->route('media.index')
-            ->with('success','Media added successfully');
+            return redirect()->route('media.index')
+                ->with('success', 'Media added successfully');
+        } else
+            return view('errors.permission');
     }
 
     /**
@@ -93,8 +103,11 @@ class MediaController extends Controller
      */
     public function show($id)
     {
-        $media = Media::find($id);
-        return view('media.show',compact('media'));
+        if (Auth::check()) {
+            $media = Media::find($id);
+            return view('media.show', compact('media'));
+        } else
+            return view('errors.permission');
     }
 
     /**
@@ -105,8 +118,11 @@ class MediaController extends Controller
      */
     public function edit($id)
     {
-        $media = Media::find($id);
-        return view('media.edit',compact('media'));
+        if (Auth::check()) {
+            $media = Media::find($id);
+            return view('media.edit', compact('media'));
+        } else
+            return view('errors.permission');
     }
 
     /**
@@ -118,16 +134,19 @@ class MediaController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'source' => 'required',
-            'address' => 'required',
-            'type_id' => 'required',
-            'user_id' => 'required',
-        ]);
+        if (Auth::check()) {
+            $this->validate($request, [
+                'source' => 'required',
+                'address' => 'required',
+                'type_id' => 'required',
+                'user_id' => 'required',
+            ]);
 
-        Media::find($id)->update($request->all());
-        return redirect()->route('media.index')
-            ->with('success','Media updated successfully');
+            Media::find($id)->update($request->all());
+            return redirect()->route('media.index')
+                ->with('success', 'Media updated successfully');
+        } else
+            return view('errors.permission');
     }
 
     /**
@@ -138,8 +157,11 @@ class MediaController extends Controller
      */
     public function destroy($id)
     {
-        Media::find($id)->delete();
-        return redirect()->route('media.index')
-            ->with('success','Media deleted successfully');
+        if (Auth::check()) {
+            Media::find($id)->delete();
+            return redirect()->route('media.index')
+                ->with('success', 'Media deleted successfully');
+        } else
+            return view('errors.permission');
     }
 }
